pyproject.toml
[project]
name = "netty"
version = "0.1.0"
description = "Netty network automation tool"
authors = [
    { name = "wangxin.jeffry", email = "wangxin.jeffry@gmail.com" }
]
dependencies = [
    "pydantic>=2.10.5",
    "pydantic-extra-types>=2.10.2",
    "pydantic-settings>=2.7.1",
    "pyyaml>=6.0.2",
    "tzdata>=2025.1",
    "typer>=0.15.1",
    "jinja2>=3.1.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.9.3",
    "pytest>=8.3.4",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/netty"]
